Split walkabout from custom_commands Fix walkabout edge cases Closes #22
